backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy>=2.0.36  # Updated for Python 3.13 compatibility
asyncpg>=0.30.0  # Updated for Python 3.13 compatibility (pre-built wheels)
alembic>=1.14.0  # Updated for Python 3.13 compatibility
pg8000==1.30.4  # Pure Python PostgreSQL driver
# psycopg2-binary==2.9.9  # Temporarily commented - install separately if needed

# Background Jobs
celery==5.3.6
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0  # For Pydantic EmailStr

# PDF Processing
PyMuPDF>=1.24.0  # Updated for Python 3.13 compatibility (pre-built wheels)
pdf2image==1.17.0
# pytesseract==0.3.10  # Optional OCR

# AI/ML
google-generativeai==0.3.2
pillow>=10.4.0  # Updated for Python 3.13 compatibility
numpy>=2.0.0  # Updated for Python 3.13 compatibility (pre-built wheels)

# Storage
boto3==1.34.34

# Utilities
pydantic>=2.10.0  # Updated for Python 3.13 compatibility (pre-built wheels)
pydantic-settings>=2.7.0  # Updated for Python 3.13 compatibility
httpx==0.26.0
python-dateutil==2.8.2

# CSV Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Email (Brevo)
sib-api-v3-sdk>=7.6.0

# Push Notifications
pywebpush>=1.14.0
py-vapid>=1.9.0

# Monitoring (optional for now)
# sentry-sdk[fastapi]==1.40.0
